requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
DateTime==5.5
distro==1.9.0
exceptiongroup==1.3.0
Farama-Notifications==0.0.4
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
gymnasium==1.2.1
stable-baselines3==2.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.11.0
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
openai==2.3.0
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
protobuf==6.33.0
pydantic==2.12.1
pydantic_core==2.41.3
pygame==2.6.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sentry-sdk==2.42.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.14.0
torch==2.9.0
tqdm==4.67.1
tokenizers==0.22.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
vizdoom==1.3.0.dev2
wandb==0.22.2
zope.interface==8.0.1
diffusers==0.35.2
transformers==4.57.1
peft==0.17.1
huggingface_hub==0.36.0
datasets>=2.14.0
pyarrow>=12.0.0